refactor(mithril-stm): move Merkle circuit boundary to common module
