+++ mithril-stm/src/circuits/common/mod.rs
@@ -1,0 +1,3 @@
+//! Shared circuit-facing boundary types and adapters.
+
+pub(crate) mod merkle;
